Move compute_all_general_metrics out of basic_metrics.py.
All the same code is being called in all the same places, this PR is just moving it into its own file. My plan is in the next PR, I will try to simplify the call stack for this method.
